requirements.txt
python-dotenv==1.0.0
pydantic==2.5.3
loguru==0.7.2
tqdm==4.66.1

sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
transformers>=4.36.2
torch>=2.1.2

pandas==2.1.4
numpy>=1.26.3
pyarrow>=14.0.2
scikit-learn>=1.4.0
scipy>=1.12.0
rank-bm25==0.2.2

huggingface_hub
datasets>=2.18

matplotlib>=3.8.2
seaborn>=0.13.1

pytest==7.4.4